Got v1 users/tracks endpoint working
