+++ discovery-provider/src/queries/queries.py
@@ -434,6 +434,7 @@ def get_feed():
             else:
                 reposted_playlists = []
 
+
         if feed_filter == "original":
             tracks_to_process = created_tracks
             playlists_to_process = created_playlists
